Add debug mode to event collector
This adds a new `debug` flag which is present in the `events.json`
output if a custom date was used as the first argument to `collect.py`.

The purpose of this is so that the frontend can show the correct time on
the clock when debugging events from a past date.

A bug with recurring events has been fixed. Instead of incorrectly
modifying the dates of recurring event templates, we now use the
`singleItems` flag to expand recurring events into individual events.

Finally, the output JSON was updated so that keys are now camel case
instead of snake case, and an `id` field is included for each event.